server/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
boto3==1.35.83
botocore==1.35.83
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
Flask==3.1.0
Flask-SocketIO==5.4.1
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
MarkupSafe==3.0.2
multidict==6.1.0
openai==1.58.1
packaging==24.2
pillow==11.0.0
postgrest==0.18.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.1
python-socketio==5.11.4
realtime==2.0.6
redis==5.2.1
requests==2.32.3
s3transfer==0.10.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
